+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/chat/dto/response/ChatRoomCreateResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class ChatRoomCreateResponse {
-    private Long chatRoomId;
+    private Integer chatRoomId;
     
     public static ChatRoomCreateResponse from(ChatRoom chatRoom) {
         return ChatRoomCreateResponse.builder()
